pyproject.toml
[project]
name = "plant-clef-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "kornia>=0.8.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "timm>=1.0.15",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "wandb>=0.19.9",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "kaggle>=1.7.4.2",
    "pre-commit>=4.1.0",
]

[tool.ruff]
# Set the maximum line length to 79.
line-length = 100
